// src/Model/iTechRestApi/iTechDataWebFilterEnum.ts
//*************************DO NOT MODIFY**************************
//
//THESE FILES ARE AUTOGENERATED WITH A TYPESCRIPT CODE GENERATION TOOL AND ANY MODIFICATIONS MADE HERE WILL BE LOST.
//PLEASE VISIT http://frhagn.github.io/Typewriter/ TO LEARN MORE ABOUT THIS VISUAL STUDIO EXTENSION
//
//*************************DO NOT MODIFY**************************



export enum iTechDataWebFilterEnum {
    advanced = 1,
    tree = 2,
    graph = 3,
    virtualTable = 4,
    rule = 5
}